+++ packages/plugin-file-upload/web/src/consts.js
@@ -1,3 +1,4 @@
+/* eslint-disable camelcase */
 export const DEFAULTS = {
   config: {},
   configViewer: {
@@ -11,4 +12,22 @@ export const DEFAULTS = {
       ),
     downloadTarget: '_blank',
   },
+};
+
+export const THEME = colors => {
+  const { textColor, secondaryColor } = colors;
+  return {
+    file_upload_name: {
+      color: textColor,
+    },
+    file_upload_type: {
+      color: textColor,
+    },
+    file_upload_icon: {
+      color: textColor,
+    },
+    file_upload_container: {
+      borderColor: `${secondaryColor} !important`,
+    },
+  };
 };