+++ packages/wrapper/web/src/themeStrategy/themes/viewer.js
@@ -1,0 +1,10 @@
+/* eslint-disable camelcase */
+export default function viewer(colors) {
+  const { actionColor } = colors;
+  return {
+    quote: {
+      'border-left-color': actionColor,
+      'border-right-color': actionColor,
+    },
+  };
+}